requirements.txt
asgiref==3.9.2
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==2.0.0
channels==4.3.1
constantly==23.10.4
cryptography==46.0.1
daphne==4.2.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.5.2
numpy==2.3.3
pandas==2.3.2
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.2
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
zope.interface==8.0